refactor: Store - invitation 경로 수정
