+++ src/components/invitation/create/InvitationInfoContainer.tsx
@@ -6,12 +6,12 @@ import InvitationInfo from '@/components/invitation/create/InvitationInfo';
 import InvitationVisitorsList from '@/components/invitation/create/InvitationVisitorsList';
 import CREATE_TEXTS from '@/constants/invitation/createTexts';
 import COMPONENT_NAME from '@/constants/common/pages';
-import useViewStore from '@/stores/usePagesStore';
-import useModalStore from '@/stores/useModalStore';
-import useSaveStore from '@/stores/useSaveStore';
-import useAlertStore from '@/stores/useAlertStore';
-import useBottomSheetStore from '@/stores/useBottomSheetStore';
-import useInvitationCreateStore from '@/stores/useInvitationCreateStore';
+import useViewStore from '@/stores/common/usePagesStore';
+import useModalStore from '@/stores/common/useModalStore';
+import useSaveStore from '@/stores/common/useSaveStore';
+import useAlertStore from '@/stores/common/useAlertStore';
+import useBottomSheetStore from '@/stores/common/useBottomSheetStore';
+import useInvitationCreateStore from '@/stores/invitaion/useInvitationCreateStore';
 import { css } from '@emotion/react';
 import { useEffect, useState } from 'react';
 import { postInvitation } from '@/pages/api/invitation/createRequests';
