+++ src/components/invitation/create/InvitationPlace.tsx
@@ -1,8 +1,8 @@
 import Button from '@/components/common/Button';
 import RadioBtn from '@/components/common/RadioBtn';
 import CREATE_TEXTS from '@/constants/invitation/createTexts';
-import useBottomSheetStore from '@/stores/useBottomSheetStore';
-import useInvitationCreateStore from '@/stores/useInvitationCreateStore';
+import useBottomSheetStore from '@/stores/common/useBottomSheetStore';
+import useInvitationCreateStore from '@/stores/invitaion/useInvitationCreateStore';
 import getIsCommonData from '@/utils/getIsCommonData';
 import getPlaceId from '@/utils/getPlaceId';
 import theme from '@/styles/theme';
