+++ src/components/invitation/create/InvitationPurposeContainer.tsx
@@ -5,9 +5,9 @@ import Category from '@/components/common/Category';
 import Button from '@/components/common/Button';
 import CREATE_TEXTS from '@/constants/invitation/createTexts';
 import COMPONENT_NAME from '@/constants/common/pages';
-import useViewStore from '@/stores/usePagesStore';
-import useInvitationHeaderTitleStore from '@/stores/useInvitationHeaderTitleStore';
-import useInvitationCreateStore from '@/stores/useInvitationCreateStore';
+import useViewStore from '@/stores/common/usePagesStore';
+import useInvitationHeaderTitleStore from '@/stores/invitaion/useInvitationHeaderTitleStore';
+import useInvitationCreateStore from '@/stores/invitaion/useInvitationCreateStore';
 import { useEffect, useState } from 'react';
 import { css } from '@emotion/react';
 import { COMMON_CATEGORIES } from '@/constants/common';
