+++ src/components/invitation/create/InvitationVisitorsContainer.tsx
@@ -4,11 +4,11 @@ import Alert from '@/components/common/Alert';
 import Add from '@/components/common/Add';
 import AddressBook from '@/components/common/AddressBook';
 import InvitationVisitorsList from '@/components/invitation/create/InvitationVisitorsList';
-import useAlertStore from '@/stores/useAlertStore';
-import useViewStore from '@/stores/usePagesStore';
+import useAlertStore from '@/stores/common/useAlertStore';
+import useViewStore from '@/stores/common/usePagesStore';
 import CREATE_TEXTS from '@/constants/invitation/createTexts';
 import COMPONENT_NAME from '@/constants/common/pages';
-import useInvitationCreateStore from '@/stores/useInvitationCreateStore';
+import useInvitationCreateStore from '@/stores/invitaion/useInvitationCreateStore';
 import theme from '@/styles/theme';
 import { css } from '@emotion/react';
 import { useState } from 'react';
