+++ src/components/invitation/create/InvitationVisitorsList.tsx
@@ -1,9 +1,9 @@
 import Add from '@/components/common/Add';
 import NameTag from '@/components/common/NameTag';
 import BottomSheet from '@/components/common/BottomSheet';
-import useViewStore from '@/stores/usePagesStore';
-import useBottomSheetStore from '@/stores/useBottomSheetStore';
-import useInvitationCreateStore from '@/stores/useInvitationCreateStore';
+import useViewStore from '@/stores/common/usePagesStore';
+import useBottomSheetStore from '@/stores/common/useBottomSheetStore';
+import useInvitationCreateStore from '@/stores/invitaion/useInvitationCreateStore';
 import CREATE_TEXTS from '@/constants/invitation/createTexts';
 import InvitationAddVisitorList from '@/components/invitation/edit/InvitationAddVisitorList';
 import theme from '@/styles/theme';
